# py36+ tests are definition ordered [v2]