--- a/changelog/5196.feature.rst
+++ b/changelog/5196.feature.rst
@@ -1,3 +1,3 @@
 Tests are now ordered by definition order in more cases.
 
-In a class hierarchy, tests from base classes are now consistently ordered before tests defined on thier subclasses (reverse MRO order).
+In a class hierarchy, tests from base classes are now consistently ordered before tests defined on their subclasses (reverse MRO order).